requirements/base.txt
# Base requirements - common to all environments

# Django
Django>=5.0,<6.1

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# API Documentation
drf-spectacular>=0.27.0

# Filtering
django-filter>=23.5

# CORS headers
django-cors-headers>=4.3.0

# Environment variables
python-dotenv>=1.0.0

# Timezone support
pytz>=2024.1

# Cloudinary for file storage
cloudinary>=1.36.0

# AI Integration
groq>=0.4.0

# Email (Resend)
resend>=2.0.0

# Redis (for caching and rate limiting)
django-redis>=5.4.0

# Celery for async task processing
celery[redis]>=5.3.0

# Celery Beat (scheduled tasks)
django-celery-beat>=2.6.0

# Two-Factor Authentication
pyotp>=2.9.0
qrcode[pil]>=7.4.0

# HTTP client for webhooks
httpx>=0.27.0
